Create v1 app for REST API
I've decided to fully isolate the v1 REST API into its own FastAPI app
to present an isolated API docs page. I've also co-located the models
with the app so that it'll be easier to create new response and request
models for other versions of the API down the road.