Mock up the HTML endpoint
This adds a render_html method to the Page domain that converts a
notebook into HTML; this could be further customized by building a
custom nbconvert HTMLExporter, adding nbconvert configurations, and/or
using a custom template.

The /html endpoint uses the template-rendered notebook, but not a
computed notebook. This is done to provide useful functionality for
testing the front-end while the interface between Times Square and
Noteburst is worked out.