Implement GET /v1/pages
We want to optimize this endpoint around the task of getting summary
info about available pages. We don't need the full page metadata from
this endpoint, so to make it faster we're getting a subset of data from
the database (only the name so far) via the PageStore into the
PageSummaryModel domain model (a subset of PageModel).

The endpoint itself should eventually support sorting and filtering, and
possibly pagination if needed. For now this provides the MVP to support
times-square-ui.