Future refactor redis stores
Now the redis stores are direct subclasses of RedisStore, so there's no
longer any facade over the common methods.

To make this possible, I created a new domain model, PageInstanceId,
that combines a page's name and value attributes. These two parameters
are commonly used together to identify an *instance* of a page, that is,
a rendering of a given page with a set of values.