Implement a /pages/:page/htmlstatus endpoint
This endpoint provides info to clients about the HTML product without
triggering a new HTML render. Clients are expected to poll this endpoint
while waiting for HTML to become available, and also to get the
latest version of an HTML page that has been re-renderered.

The "available" field can tell clients whether the HTML exists yet.

The "html_hash" field is a SHA256 has that provides a version stamp for
the HTML content.