Have /htmlstatus use PageService.get_html
It turns out in the Times Square UI model that we really do need the
/htmlstatus endpoint to process a notebook from a noteburst job result.
And conceivably if a noteburst job is lost, to also trigger a new
noteburst job. Otherwise the SWR request in Times Square UI to
/htmlstatus would never process the noteburst result in the same way
that a request to /html would.

Thus we can drop the PageService.get_html_status method entirely and use
get_html for both endpoints. The only difference is the response
format/payload.