Install safir with arq from DM-34473
Note that we need to use a re-release version of arq 0.23 to
compatibility with current aioredis. pip-tools compile adds the arq
redis extra, which actually breaks the installation from
requirements.txt; I need to manually remove this annotation.
Hopefully this will go away when arq is out from
pre-release and-or we're installing safir from pypi again.