Add GitHub repository domain models.
These domain models describe a checkout of a GitHub repository, the
Times Square settings and sidecar YAML configuration files, and also the
GitHub API responses for the Git tree and blob APIs.

This code *might* be in the wrong place since it's describing external
domains/interfaces (settings files and GitHub API results). On the other
hand, GitHubRepositoryCheckout acts as an internal domain model for
services that ingest content from GitHub.

tl;dr We might need to shuffle this code around once we have internal
models for a GitHub repository as it impacts TS's information
architecture, not temporary checkouts during an ingest service workflow.