Add support for the installation webhook
We previously supported the installation_repositories event, but it's
the "installation" event that's triggered when the Times Square GitHub
App is initially installed; we need to use this event to initially seed
repositories (or delete repositories for uninstallation). The
installation_repositories event is then used to incrementally add more
repos to existing installations.