Implement /v1/github endpoint
This endpoint provides the inventory of GitHub-backed pages;
hierarchically organized. Squareone uses this data to generate the
navigation UI for the GitHub section of Times Square's UI.

The new githubtree domain deals with the organization of GitHub-backed
pages within the owner/repo/directory/page hierarchy.

I haven't measured the performance of algorithm that converts the linear
listing of pages from SQL into the hierarchical structure; however since
this data only changes when the GitHub sources are re-synced, this data
is a candidate for aggressive caching.