Introduce "display_settings" for HTML renders
NbHtmlModels are now rendered with display parameters
(NbDisplaySettings) that affect how the HTML is rendered, and are
conceptually separate from parameters, which affect the Jupyter Notebook
computation.

The PageService now creates a full matrix of HTML renders for the
different display settings. At the moment the only setting is to
show/hide code input cells. In the future, a theme option to choose a
light or dark rendering might get added.

The URL query string for toggling the show/hide code cells behaviour is
ts_hide_code (1 or 0).