Push redis key generation into PageInstanceId
Before the RedisBase was calculating the redis key from PageInstanceId.
Then we started using a subclass of PageIstanceId called NbHtmlKey
specifically for HTML caching. But since the key code was in the
RedisBase, we didn't realize we needed to update the code for
calculating the key (NbHtmlKey contains display settings).