Add GitHubRepositoryCheckout.create
This classmethod makes it possible for a GitHubRepositoryCheckout to
create itself from a github client, repository, and sha. This eliminates
need for the "create_checkout" method on the RepositoryService. Overall
this means it's easier to create checkouts from different parts of the
code, which will be useful when writing the check run validation code.