Move check run setup into domain classmethods
Rather than having extensive logic for setting up a check run in the
service layer, now class methods on the GitHubConfigsCheck are able to
create and update the status of check runs. Overall the intent of this
work is to make it easier to add check runs for notebooks in the next
phase of work, and reduce the amount of complexity spread into the
service layer.