Implement github check for notebook execution
This is a first pass at running a github check for notebook execution.

One thing that's a bit awkward is making PageService.add_page return a
new PageExecutionInfo instance so the check run has access to the
noteburst job corresponding to the page; it may be better to separate
adding a page to the database from executing the notebook so that
PageExecutionInfo doesn't need to have lots of nullable types.