Add service and handers for getting PR-based pages
This provides access to the pages that are built automatically by the
GitHub Check run handlers for a GitHub pull request.

- New PageService.get_github_pr_page
- New /github-pr/ endpoints, similar to the existing /github/ endpoints,
  but have the PR's head commit SHA in the path to show pages for a
  specific PR check run
- Revised GitHubTree domain model.

The new domain model separates path segments (which are used to add
nodes relative to other nodes) from the Squareone URL path. This
dramatically simplifies the tree domain construction and makes it
possible to use the same domain object for PR preview trees as well.

The new GitHubNode is a dataclass; in the REST API we transform that
dataclass into a Pydantic model for the response.