Filter webhook events based on accepted_github_orgs
This is done with the filter_installation_owner decorator that's applied
to all webhook handlers. I added tests for this decorator with a mock
github API client.

To make the decorator work, I needed to add the github webhook factory
to the handler call pattern so that the decorator can make a call to the
github API to check where the request came from.

Eventually it'd be smart to cache the association of installation ID to
an owner to reduce the number of requests to GitHub.

I'm only apply the decorator to handlers that add data to Times Square
or cause new computations. Events that remove a repository will always
get processed so we don't accidentally keep around a repository we
wanted to get rid of.