Replace first code cell with parameter template
This implements a new default behaviour in Times Square: the first code
cell is now intended to only set parameters in a notebook. When Times
Square renders out a notebook with parameter values, it replaces that
first code cell with totally new content that sets Python variables to
the chosen values.

Previously all code cells were treated as templates, but this meant that it
was hard to test notebook interactively in JupyterLab because the
notebook can't run code with unrendered Jinja syntax.