Add DELETE /pages/:page/html
This endpoint allows for soft-deletion of the cached HTML for a page
instance. The service launches a noteburst execution of the notebook and
launches a replace_nbhtml background task to monitor that job and
ultimately replace the cached HTML for the page.

This is different from regular HTML creation triggered by a GET request
in that the existing HTML remains available to clients while the new
page is being rendered.