Add GET /v1/pages/:page/html/events SSE endpoint
This is a server-sent events endpoint that provides a stream of updates
about a notebook's execution. This is particularly useful for a client
to follow the progress of recalculating a page instance. This endpoint
might also be able to replace the /htmlstatus endpoint so that a client
no longer needs to poll it for the current HTML version.

Right now the event generator implements its own polling loop to ask
noteburst for status while a job is in progress. Ideally we'd like to
only send an event when there's a change (i.e. noteburst API sends its
own SSE stream driven by the arq queue events) and that extra events
aren't sent while the HTML remains static to reduce the overall
bandwidth.