Set arq timeouts for check runs
We want these check runs to be slightly longer than the configuration so
that the worker's code has a change to report the timeout before arq
kills the job itself and we lose the opportunity to report the check run
as failed to GitHub.

It seems that there are typing issues with using arq.worker.func, so
we'll try to ignore those for now.