Add new TS_DEFAULT_EXECUTION_TIMEOUT env var
This configuration sets an app-wide default configuration for the
timeout to set when executing a notebook with noteburst (either during a
check run for a PR or in actual use). Using a check run should help
ensure that we don't have hanging executions or check runs and can also
encourage authors to keep pages fast.

In the future we'll add per-page customizations to the timeout so
authors can set longer timeouts when needed, or even set shorter
timeouts to fail fast in quick notebooks.