Create PageParameters class for parameter domain
Previously a Page domain model stored parameter schemas as a dict of
name and PageParameterSchema. We're replacing that dict with a class so
that methods related to parameters on the Page model can now be
co-located in this new PageParameters class. This helps reduce the size
of the page domain model and will make adding to the parameters domain
easier in the future.