Refactor code templating methods to PageParameters
Move code related to serializing parameter values to strings and to
creating the template for the parameters code cell into the
PageParameters class. Again, this helps to simplify the top-level
complexity of the PageModel.

We might consider moving the render_parameters method onto the
PageInstanceModel and move all code related to a page with specific
values to that class.