Make type-specific parameter schema classes
This treats PageParameterSchema as an abstract base class that's
implemented by a subclass specific to each type as indicated by the
schema.

Now the PageParameterSchema.create factory creates these type-specific
objects. It turns out to have little impact on usage of this API -- it
is a very transparent change because of type compatibility.

The advantage of doing this is that now the cast method can be much
shorter and focused on the type. This makes it easier to support more
types in the future.