Switch code cell construction from Jinja
Rather than use Jinja to create the Python code cell, we generate it
directly in code. This gives us more flexibility in the assignments that
we'll be able to use to render out dates and datetime types.

The PageParameterSchema types have a new method,
create_python_assignment, that supports creating this code cell and can
be tuned to each parameter type.