Move ipynb template rendering to PageInstance
Previously the Page's templated ipynb file was rendered with
Page.render_parameters. Here we move that functionality to
PageInstanceModel.render_ipynb. This is useful because the
PageInstanceModel is created with resolved and validated parameter
values so we know that render_ipynb will be performed with those good
values. Previously each call to PageModel.render_parameters needed to
resolve and validate those values first. But now the PageInstanceModel
provides a useful place to encapsulate those resolved values.