Separate PageInstanceModel and PageInstanceIdModel
Previously PageInstanceModel inherited from PageInstanceIdModel. But
this makes it hard to change how the cache_key is computed in the future
because we'd like to start using PageParameters to help compute the
cache_key, but those aren't available from PageInstanceIdModel. The
solution is to make PageInstanceModel stand alone, and provide an `id`
property that generates a "static" PageInstanceIdModel.

Also added a PageInstanceModel.page_name attribute since the
PageIdModel.name attribute is no longer available.