Compute URL query strings via PageParameterSchemas
Previously when we generated a URL query or a Redis key for the
NbHtmlCacheStore or NoteburstJobStore, we assumed that we could cast the
parameter values directly to str. But this is limiting when we introduce
complex types like date and datetime strings. Here, we want to use the
PageParameterSchema as the source of truth for how their values should
be serialized into a query-string compatible type.

To accomplish that we've added create_qs_value on PageParameterSchema
classes that handles the serialization of a value into a string
representation that's expected for a URL query string.

At the same time, we used this method to generate the cache key for both
PageInstanceIdModel and NbHtmlKey. These new cache keys are now reliable
with complex types for parameters. Note that this does break the cache
keys of existing stores.