Compute URL qs of /htmlstatus from NbHtmlKey
The /htmlstatus endpoint seems to be the last place where we were adhoc
computing the canonical query string for a page instance's URL. This
changes that URL computation to use NbHtmlKey's url_query_string
property.

To do this, I changed the /htmlstatus endpoint so that it doesn't use
the get_html service, but rather a new get_html_and_status service that
returns a new NbHtmlStatusModel that contains both the NbHtmlModel and
the NbHtmlKey model. The existing get_html service uses that as its
basis.