Use PageInstanceIdProtocol
Previously the NbHtmlKey inherited from PageInstanceIdModel. This was
useful for the interface typing of the RedisPageInstanceStore, but had
the downside of make it unusually buggy to compute the cache key and url
query string. Now we're switching to a model of composing the
PageInstanceId _into_ the NbHtmlKey so it's clearer how the components
of the cache key and url query string are being calculated. To continue
making these classes work as identifiers for RedisPageInstanceStore,
we're introducing a new PageInstanceIdProtocol that indicates both
NbHtmlKey and PageInstanceIdModel must provide attributes cache_key and
url_query_string.