Add support for a date parameter type
This adds support for date parameers (e.g. "2025-02-21"). In JSON
Schema, dates are string types with a "format" key set to "date".
Therefore in URL query strings we expect dates to be ISO8601 strings.

However in the parameter value assignment cell we automatically cast
that string to a native python datetime.date type. To do this, we've
added the capability for each PageParameterModel to declare what import
statements need to be made for parameter assignments. Then
PageInstanceModel._create_parameter_assignment_cell adds those import
statements to the parameters assignment cell.

For testing, added a date parameter to the demo notebook.