Implement migrate-html-cache CLI command
This CLI command can migrate the NbHtmlCache keys from the old format
(based on a b64 encoding of a JSON object with the parameter values and
display settings) to the new format computed more reliably from the URL
query string.

This command can run in a dry-run mode and can also work on either a
specific page, or with all pages.