Use post init hook with PageInstanceModel
Drop the create classmethod and instead do the parameter validation in a
post init hook. This ensures that parameter values are always resolved.
The downside is that since we treat query strings as having a Mapping
type, we need to cast those values to a dict when creating the
PageInstanceModel.