Run nbstripout when creating a new page
The strip_ipynb method strips the ipynb source in-place using
nbstripout. We call this in the two entrypoints for creating a new
PageModel: the classmethods for creating a page from a github repo, and
from an api upload. This placement of the strip_output process should
ensure that any new/update notebooks are stripped, but we don't
unnecessarily strip notebooks each time we create a PageModel (e.g.,
from a database).

In terms of what we strip: we strip all outputs, and the kernelspec
metadata.