Replace dynamic default with default in API resp
For a page with a parameter that has a dynamic default, the API now gets
the "strict_schema" rather than the original JSON schema for the
parameter. With the strict schema, we remove the X-Dynamic-Default
extension and replace it with a default value that is computed based on
the parameter's rules for calculating the default at that moment.

This presents pages with dynamic defaults as being the same as a page
with a static default, thereby making dynamic defaults a concern of the
page author, rather than the page viewer.