Add support for a dayobs parameter format
The new "dayobs" parameter type adds native support support for Rubin
Observatory's "dayobs" date format. The dayobs is a string, `YYYYMMDD`,
that corresponds to the date in the UTC-12 timezone. This convention is
useful because it is a unique identifier for each observing night on
Cerro Pachón, and it is used in the Rubin Observatory's data products.
To use dayobs, create a parameter of type `string` with a format field
set to `dayobs`.

In the sidecar YAML schema, the format key can be set to "dayobs".
Internally it is converted to JSON schema with the X-TS-Format key set
to dayobs since dayobs isn't a standard format.

When we export the schema to the public API, we drop the X-TS-Format key
but add a basic regex pattern for client-side validation.