Fix handling of relative days/weeks/yeares
The documented syntax of +1week/+1month/+1year from the 0.21.0 release
notes weren't actually implemented correctly. We've also taken the
opportunity to fix the consistency with how relative numbers of days are
computed (+/-Nd), so the new syntax is +1w, -2m, +5y, etc. We don't have
any users of this feature from the 0.21.0 release, so this is a good
opportunity to fix it.