Use json serialization for storing rruleset
We've discovered that dateutil.rrule.rruleset's __str__ method does not
actually round-trip with the rrulesetstr method! The __str__ method
drops timezone information, which means that the rruleset created by
rruleststr can't be compared with a timezone-aware time window.

So our solution is to use our own JSON serialization of the rruleset and
store that in the database.

The ExecutionSchedule domain object now takes that serialized JSON.

We now treat ScheduleRule as a domain object, and the custom root
ScheduleRruleset is now responsible for serialization and
deserialization. This isn't perfect because we use the same pydantic
model in the public interface for the sidecar file format... we should
separate the internal domain and database model from the public
interface. We'll address that next.