Redesign the schema for schedule rules
The Pydantic model for the schedule based on Semaphore was just too hard
to work with because rrule is inherently flexible. To make scheduling
easier to reason about, document, and ultimately model with Pydantic, I
decided to break up the pydantic model for a rule into three:

- ScheduleRule - for all the complex rules based on
  weekday/hour/week/day_of_year and the like. Takes freq as a frequency
  baseline
- ScheduleFromDate - takes a start date, and then repeats at the given
  frequency and interval
- ScheduleDate - schedules execution on a specific date

I've taken out timezone handling, and defaulting to UTC for
ScheduleRule, though ScheduleFromDate and ScheduleDate can use other
timezones based on the datetime string.