Add dayobs-date format for nicer dayobs parameters
dayobs follows the strict YYYYMMDD format because formally it is an
integer, not a string, input in notebooks.

We've introduced a related foramt here, dayobs-date, that uses the same
timezone spec as dayobs, but takes YYYY-MM-DD strings as parameter
values, and in notebooks, provides datetime.date instances that are easy
to use.