Change dayobs format to create ints in Python
After discussion with early users, we realized that dayobs as it is
formally designed in https://sitcomtn-032.lsst.io is actually an
integer, not a string! Since we have dayobs-date now as a user-friendly
format, we can let dayobs be closer to the spec.

This is a breaking change, but there should be relatively few users
because the last version was announced only a week ago.